Refactor: splits code to simplify tests
